% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{Arrange}
\alias{Arrange}
\alias{Arrange,Distribution-method}
\title{Arrange a Distribution}
\usage{
\S4method{Arrange}{Distribution}(x)
}
\arguments{
\item{x}{A \code{\link[=Distribution-class]{Distribution}}.}
}
\value{
A \code{\link[=Distribution-class]{Distribution}} that is like
  \code{x}, except that its first \code{n-1} columns are arranged in
  alphabetical order, and its rows are arranged such that the values of the
  alphabetized columns are in increasing order.
}
\description{
Arrange a Distribution by alphabetizing its first \code{n-1} columns and
by arranging the values of these alphabetized columns, in increasing order.
}
