% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{as_tibble.Distribution}
\alias{as_tibble.Distribution}
\title{Coerce a Distribution to a Tibble}
\usage{
\method{as_tibble}{Distribution}(x)
}
\arguments{
\item{x}{A \code{\link[=Distribution-class]{Distribution}}.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Coerce a distribution to a tibble. This function is used internally in many
other functions in order to capitalize on functionality from other packages
such as \code{dplyr}.
}
