% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Bhattacharya}
\alias{Bhattacharya}
\alias{Bhattacharya1}
\alias{Bhattacharya1,Distribution,Distribution-method}
\alias{Bhattacharya2}
\alias{Bhattacharya2,Distribution,Distribution-method}
\title{Bhattacharya Distances}
\usage{
\S4method{Bhattacharya1}{Distribution,Distribution}(p1, p2)

\S4method{Bhattacharya2}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
A Bhattacharya distance between distributions \code{p1} and \code{p2}.
}
\description{
Compute the Bhattacharya distance 1 or the Bhattacharya distance 2 between
two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space. Let \code{rho(p1, p2)}
denote the fidelity similarity between \code{p1} and \code{p2}. Then, the
Bhattacharya distance 1 between \code{p1} and \code{p2} is given by
\code{acos(rho(p1, p2))^2}; the Bhattacharya distance 2 between \code{p1}
and \code{p2} is given by \code{-log(rho(p1, p2))}.
}
\seealso{
\code{\link{Fidelity}}
}
