% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{Distribution-class}
\alias{Distribution-class}
\title{S4 Class: Distribution}
\description{
The \code{Distribution} class represents discrete probability distributions.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A data frame, with \code{n > 1} columns, that defines a
probability mass function that associates probabilities to a finite number
of values assumed by a random variable. Columns \code{1} to \code{n-1}
determine the values assumed by the random variable; none of these columns
should be named \code{"P"}. Column \code{n} defines the probability
assigned to each value assumed by the random variable; this column must be
named \code{"P"}.}
}}

