% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{Distribution}
\alias{Distribution}
\alias{Distribution,data.frame-method}
\title{Initialize a Distribution}
\usage{
\S4method{Distribution}{data.frame}(x, p = .data$P)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{p}{An unquoted column name in \code{x}, which will be renamed to
\code{P} to comply with the validity method for the Distribution class.
If the values of this column in \code{x} do not sum to 1, then it is
normalized by its sum, so that it may be treated as the values of a
probability mass function.}
}
\value{
A \code{\link[=Distribution-class]{Distribution}}
}
\description{
Initialize a distribution from a data frame or a tibble.
}
