% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Euclidean}
\alias{Euclidean}
\alias{Euclidean,Distribution,Distribution-method}
\title{Euclidean Distance}
\usage{
\S4method{Euclidean}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The Euclidean distance between distributions \code{p1} and \code{p2}.
}
\description{
Compute the Euclidean distance between two probability distributions. (Some
authors also refer to the Euclidean distance on distributions as the
Patrick-Fischer distance.)
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space. Then, the Euclidean
distance between \code{p1} and \code{p2} is given by
\code{(sum(abs(p1 - p2)^2))^(1/2)}.

The Euclidean distance is equal to the 2-metric.
}
\seealso{
\code{\link[=pMetric]{p-metric}}
}
