% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Fidelity}
\alias{Fidelity}
\alias{Fidelity,Distribution,Distribution-method}
\title{Fidelity Similarity}
\usage{
\S4method{Fidelity}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The fidelity similarity between distributions \code{p1} and \code{p2}.
}
\description{
Compute the fidelity similarity between two probability distributions. This
similarity is also known as the Bhattacharya coefficient (not to be confused
with the Bhattacharya distances) and the Hellinger affinity (not to be
confused with the Hellinger metric).
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space. Then, the fidelity
similarity between \code{p1} and \code{p2} is given by \code{sum(sqrt(p1 * p2))}.

The Hellinger metric and the Bhattacharya distances may both be defined in
terms of the fidelity similarity.
}
\seealso{
\code{\link{Bhattacharya1}}, \code{\link{Bhattacharya2}}, \code{\link{Hellinger}}, \code{\link{JeffriesMatusita}}
}
