% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Hellinger}
\alias{Hellinger}
\alias{Hellinger,Distribution,Distribution-method}
\title{Hellinger metric}
\usage{
\S4method{Hellinger}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The Hellinger metric between distributions \code{p1} and \code{p2}.
}
\description{
Compute the Hellinger metric between two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote vectors of probability mass assigned by
two distributions defined on the same state space. Let \code{rho(p1, p2)}
denote the fidelity similarity between \code{p1} and \code{p2}. Then, the
Hellinger metric between \code{p1} and \code{p2} is given by
\code{sqrt(1 - rho(p1, p2))}.

The Hellinger metric may also be expressed in terms of the Euclidean distance:
\code{(1 / sqrt(2)) * Euclidean(sqrt(p1), sqrt(p2))}.
}
\seealso{
\code{\link{Fidelity}}
}
