% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{Identical}
\alias{Identical}
\alias{Identical,Distribution,Distribution-method}
\title{Test Identity Between Distributions}
\usage{
\S4method{Identical}{Distribution,Distribution}(x, y, names = TRUE, states = TRUE, probs = TRUE, arrange = TRUE)
}
\arguments{
\item{x, y}{\code{\link[=Distribution-class]{Distributions}}.}

\item{names}{A logical. If \code{TRUE}, then the names of \code{x} and
\code{y} are compared.}

\item{states}{A logical. If \code{TRUE}, then the states of the random
variables associated with \code{x} and \code{y} (i.e., the values of the
columns other than \code{P}) are compared.}

\item{probs}{A logical. If \code{TRUE}, then the probabilities assigned by
\code{x} and \code{y} (i.e., the values of the column \code{P}) are
compared.}

\item{arrange}{A logical. If \code{TRUE}, then \code{x} and \code{y} are
respectively passed to \code{\link{Arrange}} before being compared.
Arranging allows for more lenient comparison between distributions that
may denote the same set-theoretic information that happens to be arranged
differently due to their computational implementation as data frames.}
}
\value{
A logical.
}
\description{
Test identity between various components of two distributions, e.g. the
names of the state space of the random variable, the states of the random
variable, and the probabilities assigned to the states.
}
