% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Jeffrey}
\alias{Jeffrey}
\alias{Jeffrey,Distribution,Distribution-method}
\title{Jeffrey Distance}
\usage{
\S4method{Jeffrey}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The Jeffrey distance between \code{p1} and \code{p2}.
}
\description{
Compute the Jeffrey distance between two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space; furthermore, let
these distributions be strictly positively-valued. Then, the Jeffrey distance
between \code{p1} and \code{p2} is given by
\code{KullbackLeibler(p1, p2) + KullbackLeibler(p2, p1)}.

Jeffrey distance is a symmetric function based on Kullback-Leibler divergence.
It is the sum of the Kullback-Leibler divergence of each distribution from
the other.
}
\seealso{
\code{\link[=KullbackLeibler]{Kullback-Leibler}}
}
