% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{JeffriesMatusita}
\alias{JeffriesMatusita}
\alias{JeffriesMatusita,Distribution,Distribution-method}
\title{Jeffries-Matusita Distance}
\usage{
\S4method{JeffriesMatusita}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The Jeffries-Matusita distance between distributions \code{p1} and
  \code{p2}.
}
\description{
Compute the Jeffries-Matusita distance between two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote vectors of probability mass assigned by
two distributions defined on the same state space. Let \code{rho(p1, p2)}
denote the fidelity similarity between \code{p1} and \code{p2}. Then, the
Jeffries-Matusita distance between \code{p1} and \code{p2} is given by
\code{sqrt(2 * (1 - Fidelity(p1, p2)))}.

The Jeffries-Matusita distance may also be expressed in terms of any of the
following: Euclidean distance, Bhattacharya distance 2, and Hellinger metric.
It is straightforward to arrive at any of these expressions.
}
\seealso{
\code{\link{Fidelity}}
}
