% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{JensenShannon}
\alias{JensenShannon}
\alias{JensenShannon,Distribution,Distribution-method}
\title{Jensen-Shannon Divergence}
\usage{
\S4method{JensenShannon}{Distribution,Distribution}(p1, p2, a)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}

\item{a}{A numeric constant between 0 and 1.}
}
\description{
Compute Jensen-Shannon divergences of two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space; furthermore, let
these distributions be strictly positively-valued. Let \code{p3 = a*p1 + (1-a)*p2},
for a constant \code{0 < a < 1}. Then, the Jensen-Shannon divergence is
given by \code{a*KullbackLeibler(p1, p3) + (1-a)*KullbackLeibler(p2, p3)}.

Jensen-Shannon divergence is a symmetric function only if \code{a = 0.5}.
}
\seealso{
\code{\link[=KullbackLeibler]{Kullback-Leibler}}, \code{\link{Topsoe}}.
}
