% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Manhattan}
\alias{Manhattan}
\alias{Manhattan,Distribution,Distribution-method}
\title{Manhattan Distance}
\usage{
\S4method{Manhattan}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The Manhattan distance between distributions \code{p1} and \code{p2}.
}
\description{
Compute the Manhattan distance between two probability distributions. This
distance is also known as the taxicab distance, the city block distance,
the rectilinear distance, and the snake distance.
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space. Then, the Manhattan
distance between \code{p1} and \code{p2} is given by \code{sum(abs(p1 - p2))}.

The Manhattan distance is equal to the 1-metric. It is also equal to twice
the trace distance.
}
\seealso{
\code{\link[=pMetric]{p-metric}}, \code{\link{TotalVariation}}
}
