% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{Topsoe}
\alias{Topsoe}
\alias{Topsoe,Distribution,Distribution-method}
\title{Topsoe Distance}
\usage{
\S4method{Topsoe}{Distribution,Distribution}(p1, p2, metric = FALSE)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}

\item{metric}{A logical. If \code{TRUE}, then the square root of the Topsoe
distance is returned, which satisfies the properties of a metric.}
}
\value{
The Topsoe distance or the Topsoe metric (if \code{metric = TRUE})
  between \code{p1} and \code{p2}.
}
\description{
Compute the Topsoe distance between two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space; furthermore, let
these distributions be strictly positively-valued. Let \code{p3 = (p1 + p2) / 2}.
Then, the Topsoe distance between \code{p1} and \code{p2} is given by
\code{KullbackLeibler(p1, p3) + KullbackLeibler(p2, p3)}.

The Topsoe distance is a symmetric function based on Kullback-Leibler divergence.
It is the sum of the Kullback-Leibler divergence of the distribution
interpolated between \code{p1} and \code{p2}, from either distribution.
Equivalently, the Topsoe distance is twice the Jensen-Shannon divergence
with \code{a = 1/2}.

The Topsoe distance is not a metric (i.e., it does not satisfy the triangle
inequality); however, its square root is a metric.
}
\seealso{
\code{\link[=KullbackLeibler]{Kullback-Leibler}}, \code{\link[=JensenShannon]{Jensen-Shannon}}
}
