% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{TotalVariation}
\alias{TotalVariation}
\alias{TotalVariation,Distribution,Distribution-method}
\title{Total Variation Distance}
\usage{
\S4method{TotalVariation}{Distribution,Distribution}(p1, p2)
}
\arguments{
\item{p1, p2}{\code{\link[=Distribution-class]{Distributions}}.}
}
\value{
The total variation distance between distributions \code{p1} and \code{p2}.
}
\description{
Compute the total variation distance between two probability distributions.
}
\details{
Let \code{p1} and \code{p2} denote the vectors of probability mass assigned
by two distributions defined on the same state space. Then, the total variation
distance between \code{p1} and \code{p2} is given by \code{0.5 * sum(abs(p1 - p2))}.

The total variation distance is equal to one half of either the 1-metric or
the Manhattan distance.
}
\seealso{
\code{\link[=pMetric]{p-metric}}, \code{\link{Manhattan}}
}
